Integrate pandoc's new task_lists
Use task lists to show bullet lists with yes/no status. On non-Firefox
it is rendered through green checkmark and red cross. On Firefox it is
rendered as disabled checkboxes (pandoc's default). This is because
Firefox does not support the CSS-hack we use to hide checkboxes and show
checkmark/cross instead.

When in a .quiz class scope, task lists can now also define quizzes.